+++ Wsparcie.h
@@ -18,7 +18,7 @@ public:
     void wspieraj(Oddzial*** tab1, Wspolrzedne*** tab2) override;
     int atakuj() override;
 
-    Wsparcie(int sila_ataku, int obrona, int wytrzymalosc, int zasieg, int liczebnosc, int morale, int x, int y, Rozgrywka* rozgrywka);
+    Wsparcie(int sila_ataku, int obrona, int wytrzymalosc, int zasieg, int liczebnosc, double morale, int x, int y, Rozgrywka* rozgrywka);
 
 };
