+++ Zwarcie.h
@@ -17,7 +17,7 @@ public:
     void wspieraj() override;
     int atakuj() override;
 
-    Zwarcie(int sila_ataku, int obrona, int wytrzymalosc, int zasieg, int liczebnosc, int morale, int x, int y, Rozgrywka* rozgrywka);
+    Zwarcie(int sila_ataku, int obrona, int wytrzymalosc, int zasieg, int liczebnosc, double morale, int x, int y, Rozgrywka* rozgrywka);
 
 };
